+++ android/src/main/kotlin/com/csdcorp/speech_to_text/SpeechToTextPlugin.kt
@@ -1,9 +1,10 @@
 package com.csdcorp.speech_to_text
 
+import androidx.annotation.NonNull;
+import io.flutter.embedding.engine.plugins.FlutterPlugin
 import android.Manifest
 import android.annotation.TargetApi
 import android.app.Activity
-import android.app.Application
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Build
@@ -23,7 +24,12 @@ import io.flutter.plugin.common.PluginRegistry.Registrar
 import org.json.JSONObject
 import android.content.Context
 import android.content.BroadcastReceiver
+import android.os.Handler
+import android.os.Looper
 import android.util.Log
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.BinaryMessenger
 import org.json.JSONArray
 import java.util.*
 
@@ -50,48 +56,108 @@ enum class SpeechToTextStatus {
     available,
 }
 
+const val pluginChannelName = "plugin.csdcorp.com/speech_to_text"
+
 @TargetApi(8)
-class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
-        MethodCallHandler, RecognitionListener, PluginRegistry.RequestPermissionsResultListener {
-    private val pluginActivity: Activity = activity
-    private val channel: MethodChannel = channel
-    private val application: Application = activity.application
+/** SpeechToTextPlugin */
+public class SpeechToTextPlugin :
+        MethodCallHandler, RecognitionListener,
+        PluginRegistry.RequestPermissionsResultListener, FlutterPlugin,
+        ActivityAware {
+    private var pluginContext: Context? = null
+    private var channel: MethodChannel? = null
     private val minSdkForSpeechSupport = 21
-    private val speechToTextPermissionCode = 78521
+    private val speechToTextPermissionCode = 28521
     private val missingConfidence: Double = -1.0
     private val logTag = "SpeechToTextPlugin"
+    private var currentActivity: Activity? = null
     private var activeResult: Result? = null
     private var initializedSuccessfully: Boolean = false
     private var permissionToRecordAudio: Boolean = false
+    private var listening = false
+    private var debugLogging: Boolean = false
     private var speechRecognizer: SpeechRecognizer? = null
     private var recognizerIntent: Intent? = null
     private var previousRecognizerLang: String? = null
     private var previousPartialResults: Boolean = true
+    private var lastFinalTime: Long = 0
+    private val handler: Handler = Handler(Looper.getMainLooper())
     private val defaultLanguageTag: String = Locale.getDefault().toLanguageTag()
 
+    override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+
+        onAttachedToEngine(flutterPluginBinding.getApplicationContext(), flutterPluginBinding.getBinaryMessenger());
+    }
+
+    // This static function is optional and equivalent to onAttachedToEngine. It supports the old
+    // pre-Flutter-1.12 Android projects. You are encouraged to continue supporting
+    // plugin registration via this function while apps migrate to use the new Android APIs
+    // post-flutter-1.12 via https://flutter.dev/go/android-project-migration.
+    //
+    // It is encouraged to share logic between onAttachedToEngine and registerWith to keep
+    // them functionally equivalent. Only one of onAttachedToEngine or registerWith will be called
+    // depending on the user's project. onAttachedToEngine or registerWith must both be defined
+    // in the same class.
     companion object {
         @JvmStatic
         fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), "plugin.csdcorp.com/speech_to_text")
-            val speechPlugin = SpeechToTextPlugin(registrar.activity(), channel)
-            channel.setMethodCallHandler(speechPlugin)
+            val speechPlugin = SpeechToTextPlugin()
+            speechPlugin.currentActivity = registrar.activity()
             registrar.addRequestPermissionsResultListener(speechPlugin)
+            speechPlugin.onAttachedToEngine(registrar.context(), registrar.messenger())
         }
     }
 
-    override fun onMethodCall(call: MethodCall, result: Result) {
+    private fun onAttachedToEngine(applicationContext: Context, messenger: BinaryMessenger) {
+        this.pluginContext = applicationContext;
+        channel = MethodChannel(messenger, pluginChannelName)
+        channel?.setMethodCallHandler(this)
+    }
+
+    override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+        this.pluginContext = null;
+        channel?.setMethodCallHandler(null)
+        channel = null
+    }
+
+    override fun onDetachedFromActivity() {
+        currentActivity = null
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        currentActivity = binding.activity
+        binding.addRequestPermissionsResultListener(this)
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        currentActivity = binding.activity
+        binding.addRequestPermissionsResultListener(this)
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+        currentActivity = null
+    }
+
+    override fun onMethodCall(@NonNull call: MethodCall, @NonNull rawrResult: Result) {
+        val result = ChannelResultWrapper(rawrResult)
         try {
             when (call.method) {
                 "has_permission" -> hasPermission(result)
-                "initialize" -> initialize(result)
+                "initialize" -> {
+                    var dlog = call.argument<Boolean>("debugLogging")
+                    if (null != dlog) {
+                        debugLogging = dlog
+                    }
+                    initialize(result)
+                }
                 "listen" -> {
                     var localeId = call.argument<String>("localeId")
-                    if ( null == localeId ) {
-                      localeId = defaultLanguageTag
+                    if (null == localeId) {
+                        localeId = defaultLanguageTag
                     }
                     var partialResults = call.argument<Boolean>("partialResults")
-                    if ( null == partialResults ) {
-                      partialResults = true
+                    if (null == partialResults) {
+                        partialResults = true
                     }
                     startListening(result, localeId, partialResults)
                 }
@@ -102,10 +168,8 @@ class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
             }
         } catch (exc: Exception) {
             Log.e(logTag, "Unexpected exception", exc)
-            pluginActivity.runOnUiThread {
-                result.error(SpeechToTextErrors.unknown.name,
-                        "Unexpected exception", exc.localizedMessage)
-            }
+            result.error(SpeechToTextErrors.unknown.name,
+                    "Unexpected exception", exc.localizedMessage)
         }
     }
 
@@ -113,80 +177,104 @@ class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
         if (sdkVersionTooLow(result)) {
             return
         }
-        Log.d(logTag, "Start has_permission")
-        val hasPerm = ContextCompat.checkSelfPermission(application,
-                Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_GRANTED
-        pluginActivity.runOnUiThread { result.success(hasPerm) }
+        debugLog("Start has_permission")
+        val localContext = pluginContext
+        if (localContext != null) {
+            val hasPerm = ContextCompat.checkSelfPermission(localContext,
+                    Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_GRANTED
+            result.success(hasPerm)
+        }
     }
 
     private fun initialize(result: Result) {
         if (sdkVersionTooLow(result)) {
             return
         }
-        Log.d(logTag, "Start initialize")
+        debugLog("Start initialize")
         if (null != activeResult) {
             result.error(SpeechToTextErrors.multipleRequests.name,
                     "Only one initialize at a time", null)
             return
         }
         activeResult = result
-        initializeIfPermitted(application)
+        val localContext = pluginContext
+        initializeIfPermitted(pluginContext)
     }
 
     private fun sdkVersionTooLow(result: Result): Boolean {
         if (Build.VERSION.SDK_INT < minSdkForSpeechSupport) {
-            pluginActivity.runOnUiThread { result.success(false) }
+            result.success(false)
             return true;
         }
         return false;
     }
 
     private fun isNotInitialized(result: Result): Boolean {
-        if (!initializedSuccessfully) {
+        if (!initializedSuccessfully || null == pluginContext) {
             result.success(false)
         }
         return !initializedSuccessfully
     }
 
+    private fun isListening(): Boolean {
+        return listening
+    }
+
+    private fun isNotListening(): Boolean {
+        return !listening
+    }
+
     private fun startListening(result: Result, languageTag: String, partialResults: Boolean) {
-        if (sdkVersionTooLow(result) || isNotInitialized(result)) {
+        if (sdkVersionTooLow(result) || isNotInitialized(result) || isListening()) {
             return
         }
-        Log.d(logTag, "Start listening")
+        debugLog("Start listening")
         setupRecognizerIntent(languageTag, partialResults)
-        pluginActivity.runOnUiThread { speechRecognizer?.startListening(recognizerIntent) }
+        handler.post {
+            run {
+                speechRecognizer?.startListening(recognizerIntent)
+            }
+        }
         notifyListening(isRecording = true)
-        pluginActivity.runOnUiThread { result.success(true) }
-        Log.d(logTag, "Start listening done")
+        result.success(true)
+        debugLog("Start listening done")
     }
 
     private fun stopListening(result: Result) {
-        if (sdkVersionTooLow(result) || isNotInitialized(result)) {
+        if (sdkVersionTooLow(result) || isNotInitialized(result) || isNotListening()) {
             return
         }
-        Log.d(logTag, "Stop listening")
-        pluginActivity.runOnUiThread { speechRecognizer?.stopListening() }
+        debugLog("Stop listening")
+        handler.post {
+            run {
+                speechRecognizer?.stopListening()
+            }
+        }
         notifyListening(isRecording = false)
-        pluginActivity.runOnUiThread { result.success(true) }
-        Log.d(logTag, "Stop listening done")
+        result.success(true)
+        debugLog("Stop listening done")
     }
 
     private fun cancelListening(result: Result) {
-        if (sdkVersionTooLow(result) || isNotInitialized(result)) {
+        if (sdkVersionTooLow(result) || isNotInitialized(result) || isNotListening()) {
             return
         }
-        Log.d(logTag, "Cancel listening")
-        pluginActivity.runOnUiThread { speechRecognizer?.cancel() }
+        debugLog("Cancel listening")
+        handler.post {
+            run {
+                speechRecognizer?.cancel()
+            }
+        }
         notifyListening(isRecording = false)
-        pluginActivity.runOnUiThread { result.success(true) }
-        Log.d(logTag, "Cancel listening done")
+        result.success(true)
+        debugLog("Cancel listening done")
     }
 
     private fun locales(result: Result) {
         if (sdkVersionTooLow(result) || isNotInitialized(result)) {
             return
         }
-        var detailsIntent = RecognizerIntent.getVoiceDetailsIntent(pluginActivity)
+        var detailsIntent = RecognizerIntent.getVoiceDetailsIntent(pluginContext)
         if (null == detailsIntent) {
             detailsIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
         }
@@ -195,22 +283,27 @@ class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
                     "Could not get voice details", null)
             return
         }
-        pluginActivity.sendOrderedBroadcast(
-                detailsIntent, null, LanguageDetailsChecker(result, pluginActivity),
+        pluginContext?.sendOrderedBroadcast(
+                detailsIntent, null, LanguageDetailsChecker(result),
                 null, Activity.RESULT_OK, null, null)
     }
 
     private fun notifyListening(isRecording: Boolean) {
-        Log.d(logTag, "Notify listening")
+        debugLog("Notify listening")
+        listening = isRecording
         val status = when (isRecording) {
             true -> SpeechToTextStatus.listening.name
             false -> SpeechToTextStatus.notListening.name
         }
-        channel.invokeMethod(SpeechToTextCallbackMethods.notifyStatus.name, status)
-        Log.d(logTag, "Notify listening done")
+        channel?.invokeMethod(SpeechToTextCallbackMethods.notifyStatus.name, status)
+        debugLog("Notify listening done")
     }
 
     private fun updateResults(speechBundle: Bundle?, isFinal: Boolean) {
+        if (isDuplicateFinal( isFinal )) {
+            debugLog("Discarding duplicate final")
+            return
+        }
         val userSaid = speechBundle?.getStringArrayList(SpeechRecognizer.RESULTS_RECOGNITION)
         if (null != userSaid && userSaid.isNotEmpty()) {
             val speechResult = JSONObject()
@@ -229,85 +322,110 @@ class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
             }
             speechResult.put("alternates", alternates)
             val jsonResult = speechResult.toString()
-            channel.invokeMethod(SpeechToTextCallbackMethods.textRecognition.name,
+            debugLog("Calling results callback")
+            channel?.invokeMethod(SpeechToTextCallbackMethods.textRecognition.name,
                     jsonResult)
         }
     }
 
-    private fun initializeIfPermitted(context: Application) {
-        permissionToRecordAudio = ContextCompat.checkSelfPermission(context,
+    private fun isDuplicateFinal( isFinal: Boolean ) : Boolean {
+        if ( !isFinal ) {
+            return false
+        }
+        val delta = System.currentTimeMillis() - lastFinalTime
+        lastFinalTime = System.currentTimeMillis()
+        return delta >= 0 && delta < 100
+    }
+
+    private fun initializeIfPermitted(context: Context?) {
+        val localContext = context
+        if (null == localContext) {
+            completeInitialize()
+            return
+        }
+        permissionToRecordAudio = ContextCompat.checkSelfPermission(localContext,
                 Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_GRANTED
-        Log.d(logTag, "Checked permission")
+        debugLog("Checked permission")
         if (!permissionToRecordAudio) {
-            Log.d(logTag, "Requesting permission")
-            ActivityCompat.requestPermissions(pluginActivity,
-                    arrayOf(Manifest.permission.RECORD_AUDIO), speechToTextPermissionCode)
+            val localActivity = currentActivity
+            if (null != localActivity) {
+                debugLog("Requesting permission")
+                ActivityCompat.requestPermissions(localActivity,
+                        arrayOf(Manifest.permission.RECORD_AUDIO), speechToTextPermissionCode)
+            } else {
+                debugLog("no permission, no activity, completing")
+                completeInitialize()
+            }
         } else {
-            Log.d(logTag, "has permission, completing")
+            debugLog("has permission, completing")
             completeInitialize()
         }
-        Log.d(logTag, "leaving initializeIfPermitted")
+        debugLog("leaving initializeIfPermitted")
     }
 
     private fun completeInitialize() {
 
-        Log.d(logTag, "completeInitialize")
+        debugLog("completeInitialize")
         if (permissionToRecordAudio) {
-            Log.d(logTag, "Testing recognition availability")
-            if (!SpeechRecognizer.isRecognitionAvailable(application)) {
+            debugLog("Testing recognition availability")
+            if (!SpeechRecognizer.isRecognitionAvailable(pluginContext)) {
                 Log.e(logTag, "Speech recognition not available on this device")
-                pluginActivity.runOnUiThread {
-                    activeResult?.error(SpeechToTextErrors.recognizerNotAvailable.name,
-                            "Speech recognition not available on this device", "")
-                }
+                activeResult?.error(SpeechToTextErrors.recognizerNotAvailable.name,
+                        "Speech recognition not available on this device", "")
                 activeResult = null
                 return
             }
 
-            Log.d(logTag, "Creating recognizer")
-            speechRecognizer = createSpeechRecognizer(application.applicationContext).apply {
-                Log.d(logTag, "Setting listener")
+            debugLog("Creating recognizer")
+            speechRecognizer = createSpeechRecognizer(pluginContext).apply {
+                debugLog("Setting listener")
                 setRecognitionListener(this@SpeechToTextPlugin)
             }
             if (null == speechRecognizer) {
                 Log.e(logTag, "Speech recognizer null")
-                pluginActivity.runOnUiThread {
-                    activeResult?.error(
-                            SpeechToTextErrors.recognizerNotAvailable.name,
-                            "Speech recognizer null", "")
-                }
+                activeResult?.error(
+                        SpeechToTextErrors.recognizerNotAvailable.name,
+                        "Speech recognizer null", "")
                 activeResult = null
             }
 
-            Log.d(logTag, "before setup intent")
-            setupRecognizerIntent(defaultLanguageTag, true )
-            Log.d(logTag, "after setup intent")
+            debugLog("before setup intent")
+            setupRecognizerIntent(defaultLanguageTag, true)
+            debugLog("after setup intent")
         }
 
         initializedSuccessfully = permissionToRecordAudio
-        Log.d(logTag, "sending result")
-        pluginActivity.runOnUiThread { activeResult?.success(permissionToRecordAudio) }
-        Log.d(logTag, "leaving complete")
+        debugLog("sending result")
+        activeResult?.success(permissionToRecordAudio)
+        debugLog("leaving complete")
         activeResult = null
     }
 
     private fun setupRecognizerIntent(languageTag: String, partialResults: Boolean) {
-        Log.d(logTag, "setupRecognizerIntent")
-        if (previousRecognizerLang == null || previousRecognizerLang != languageTag || partialResults != previousPartialResults ) {
+        debugLog("setupRecognizerIntent")
+        if (previousRecognizerLang == null ||
+                previousRecognizerLang != languageTag ||
+                partialResults != previousPartialResults) {
             previousRecognizerLang = languageTag;
             previousPartialResults = partialResults
-            pluginActivity.runOnUiThread {
-                recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH).apply {
-                    Log.d(logTag, "In RecognizerIntent apply")
-                    putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM)
-                    Log.d(logTag, "put model")
-                    putExtra(RecognizerIntent.EXTRA_CALLING_PACKAGE, application.packageName)
-                    Log.d(logTag, "put package")
-                    putExtra(RecognizerIntent.EXTRA_PARTIAL_RESULTS, partialResults)
-                    Log.d(logTag, "put partial")
-                    if (languageTag != Locale.getDefault().toLanguageTag()) {
-                        putExtra(RecognizerIntent.EXTRA_LANGUAGE, languageTag);
-                        Log.d(logTag, "put languageTag")
+            handler.post {
+                run {
+                    recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH).apply {
+                        debugLog("In RecognizerIntent apply")
+                        putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM)
+                        debugLog("put model")
+                        val localContext = pluginContext
+                        if (null != localContext) {
+                            putExtra(RecognizerIntent.EXTRA_CALLING_PACKAGE,
+                                    localContext.applicationInfo.packageName)
+                        }
+                        debugLog("put package")
+                        putExtra(RecognizerIntent.EXTRA_PARTIAL_RESULTS, partialResults)
+                        debugLog("put partial")
+                        if (languageTag != Locale.getDefault().toLanguageTag()) {
+                            putExtra(RecognizerIntent.EXTRA_LANGUAGE, languageTag);
+                            debugLog("put languageTag")
+                        }
                     }
                 }
             }
@@ -350,15 +468,29 @@ class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
         sendError(errorMsg)
     }
 
+    private fun debugLog( msg: String ) {
+        if ( debugLogging ) {
+            Log.d( logTag, msg )
+        }
+    }
+
     private fun sendError(errorMsg: String) {
         val speechError = JSONObject()
         speechError.put("errorMsg", errorMsg)
         speechError.put("permanent", true)
-        channel.invokeMethod(SpeechToTextCallbackMethods.notifyError.name, speechError.toString())
+        handler.post {
+            run {
+                channel?.invokeMethod(SpeechToTextCallbackMethods.notifyError.name, speechError.toString())
+            }
+        }
     }
 
     override fun onRmsChanged(rmsdB: Float) {
-        channel.invokeMethod(SpeechToTextCallbackMethods.soundLevelChange.name, rmsdB)
+        handler.post {
+            run {
+                channel?.invokeMethod(SpeechToTextCallbackMethods.soundLevelChange.name, rmsdB)
+            }
+        }
     }
 
     override fun onReadyForSpeech(p0: Bundle?) {}
@@ -368,8 +500,7 @@ class SpeechToTextPlugin(activity: Activity, channel: MethodChannel) :
 }
 
 // See https://stackoverflow.com/questions/10538791/how-to-set-the-language-in-speech-recognition-on-android/10548680#10548680
-class LanguageDetailsChecker(flutterResult: Result, pluginActivity: Activity) : BroadcastReceiver() {
-    private val pluginActivity: Activity = pluginActivity
+class LanguageDetailsChecker(flutterResult: Result) : BroadcastReceiver() {
     private val result: Result = flutterResult
     private var supportedLanguages: List<String>? = null
 
@@ -400,12 +531,44 @@ class LanguageDetailsChecker(flutterResult: Result, pluginActivity: Activity) : 
                 localeNames.add(buildIdNameForLocale(locale))
             }
         }
-        pluginActivity.runOnUiThread { result.success(localeNames) }
+        result.success(localeNames)
 
     }
 
     private fun buildIdNameForLocale(locale: Locale): String {
         val name = locale.displayName.replace(':', ' ')
         return "${locale.language}_${locale.country}:$name"
+    }
+}
+
+private class ChannelResultWrapper(result: Result) : Result {
+    // Caller handler
+    val handler: Handler = Handler(Looper.getMainLooper())
+    val result: Result = result
+
+    // make sure to respond in the caller thread
+    override fun success(results: Any?) {
+
+        handler.post {
+            run {
+                result.success(results);
+            }
+        }
+    }
+
+    override fun error(errorCode: String?, errorMessage: String?, data: Any?) {
+        handler.post {
+            run {
+                result.error(errorCode, errorMessage, data);
+            }
+        }
+    }
+
+    override fun notImplemented() {
+        handler.post {
+            run {
+                result.notImplemented();
+            }
+        }
     }
 }